Now sending packets on pong, master / slave, now needs to interpret them and try

+++ games_sources/pong/src/PongPacket.hpp
@@ -19,8 +19,8 @@ namespace arcade
 	}		pos;
 	struct
 	{
-	  int32_t	x;
-	  int32_t	y;
+	  float		x;
+	  float		y;
 	}		ball;
 	uint8_t	id;
       };
